Apply patch 011_liga_to_dlig by Christian Perrier from Debian sources.
I am not quite sure which bug this is meant to fix, but it is most like
	bug #15792: Freefont Alef and Lamed combine
although this patch is not the patch included in that bug report.
(The report was closed...no indication if any action was taken.)

This is the last outstanding patch from the Debian sources for
	ttf-freefont
found in
	ttf-freefont-20060501cvs-12.diff
That is, the SVD files of this project should now match the ones used
to build the Debian ttf-freefont fonts.


git-svn-id: svn://svn.savannah.gnu.org/freefont/trunk/freefont@486 3f5956db-6956-4f1c-a952-c5934ae6f67f