Fixed several lookup issues, based on "Find Problems".
TamlCA_TamlSSA.half with ligature from lookup subtable
'haln' Halant Forms in Tamil lookup 24 subtable
refers to a missing glyph glyph190

Similar complaint with a 'half' lookup.

For this and the other glyph190 issue, I changed glyph190 to TamlCA_TamlSSA,
which looks right according to its surroundings.

The lookup 'aalt' Access All Alternatives i is active for glyph bn_ekaar
which has script 'bng2', yet this script does not appear in any of the
features which apply the lookup.

Similar complaint about bn_aikaar.

For this I changed the script of the lookup to bng2, and changed the name of
the lookup accordingly.

The lookup 'liga' standard Ligatures look is active for glyph uni0937091F
which has script 'dev2', yet this script does not appear in any of the
features which apply the lookup.

It makes a ligature of uni0937 and uni091F, which are in Devangari.
So I changed the script of the lookup to 'dev2' and changed the name of the
lookup accordingly.

Also shortened the names of many of the lookup tables.


git-svn-id: svn://svn.savannah.gnu.org/freefont/trunk/freefont@596 3f5956db-6956-4f1c-a952-c5934ae6f67f