Changed names of a bunch of glyphs with invalid TrueType names, in response to failed Validation, and in Windows the font not showing up in the font display.
These were all in the range 0x1025f+, which is an attempt to take care
of combining forms in Indic scripts (I don't know if this work was ever
completed, or if it is the best way to do this).

The names were like
	uni0916_uni094D.half_uni0928.pres
or longer
	uni0926_uni094D.half_uni0926_uni094D.half.pres
and a few much longer than this.

By a process of removing redundant information, I think I preserved
what information was there.

There were a few more names that were never filled in, but contained
a minus, which isn't valid.  These start with "nonunicode", and I replaced
the minus with an underscore.


git-svn-id: svn://svn.savannah.gnu.org/freefont/trunk/freefont@600 3f5956db-6956-4f1c-a952-c5934ae6f67f