Overhaul of Cyrillic
Italic, BoldItalic
	added small yat for bug #22588 (note Times New Roman doesn't use
	alternate form in Italic)

All forms of Serif have big problems in Cyrillic.

The ugliest is in roman.  The letters, even of the Russian alphabet, are
of inconsistent height (awfully, small 0438 (ii) 0446 (tse))
and they vary from the height of Latin
and they vary from the height of italic and bold.
They are a mish-mash of letters from several fonts, of similar (but not
quite identical) weight, and similar, (but not quite identical) size.

I think the best solution would be to identify the face that best matches
Latin, and fill the range with that.  I think this is possible because the
rarer letters seem to be better: the common letters are the ones that
are wrong.

For now, I just increased the sized of 0438 and 0446, and 048a, 048b,
also 0459 (lje) 045A (nje) 0464 (dje)

Other issues
------------
Were outline, with home-made accents - replaced with references
	0400	(was referencing E)
	0401, 0402, 040c, 040d, 040e 0450, 0451, 0453 more in the later ranges.

Deleted dozens of characters in Bold, Italic, and BoldItalic, that appeared
half-made (a part of the letter was copied, but the letter was never
completed, and was basically wrong) Lots of letters with hooks and lines in
italic were represented as just the unhooked version in italic.

048a, 048b (short i with tail, had big-eared breve common in Cyrillic, but
	unlike other letters)

Re-implemented many compound letters with references.

Questionable
------------
In all versions, 0478-9 Uk, uk.
SerifItalic 0495

Small ge with acute, in italic: is it like italic ge, or oblique?
Well, peeked at Times New Roman, which has it oblique.


git-svn-id: svn://svn.savannah.gnu.org/freefont/trunk/freefont@668 3f5956db-6956-4f1c-a952-c5934ae6f67f