Fix for FontForge warning (v 20090923)
"On Windows many apps will have problems with this font's kerning, because
because 58 of its glyph kern pairs cannot be mapped to unicode-BMP kern pairs"

This was due to kern tables for the Musical Symbols range, which isn't in
the Unicode Basic Multilingual Plane.

Removed this lookup.  Might consider another way to position these symbols.


git-svn-id: svn://svn.savannah.gnu.org/freefont/trunk/freefont@1486 3f5956db-6956-4f1c-a952-c5934ae6f67f